Move case command parser
